build: Add pyproject.toml
This is the new "universal" metadata file for Python, since PEP 518.

The dependencies from requirements.txt were added as dev dependencies,
as well as mock, which was missing from requirements.txt. This is needed
for running the tests.

In addition, add bumpversion to the dev dependencies. For now, we stick to the
original bumpversion. At some point we should switch to
https://github.com/c4urself/bump2version.