build: Handle files correctly that aren't packages
We want to ship various files in the source distribution, but they are
not packages. Including them in the packages option of poetry seemed to
work for building the sdist and wheels, but it broke poetry install.

The correct way to do this was implemented in Poetry with
https://github.com/python-poetry/core/pull/6. This fix is in an as-yet
unreleased version, unfortunately.