feat: Allow transport objects to be closed
The base transport class now has a close method, which marks the
transport as closed and additionally will throw an exception if the call
method is called on a closed transport object. In addition, the close
method will be called by __del__, so the transport will be automatically
closed prior to it being destructed.

Transports should override the _close method if they need to free any
transport-specific resources (currently only DatabaseTransport and
SshTransport need to do so).

Fixes #64