ci: Fix poetry install errors
Poetry recommends *not* installing it using pip, since it can lead to
conflicts with your package's dependencies and Poetry's own
dependencies. This seems to be the cause of current failures.

Instead we install poetry in the recommended way, however this brings
with it its own vendored dependencies for each version of Python that it
supports and the nightly Python version may or may not have those
dependencies. For now, we simply allow nightly to fail.

Lastly, it seems that "matrix" is has been replaced by "jobs", so switch
to the current (documented) syntax.