feat: Use better names for XML attributes
The xml generator has special code to distinguish between tag names,
attributes, etc set by the itoolkit classes vs extra attributes set by
the user. This hinges on the length of the key being 1 or not, leading
to difficult to understand keys of 'i', 'k', 'j', etc. Instead, we can
use more descriptive names and prefix them with an underscore to
distinguish them.

This is the first step to refactoring this code in to more pythonic and
understandable code. The next step is to use object attributes instead
of special dictionary keys for these items.