feat: Generate XML DOM programmatically
Instead of generating an XML string and then parsing it, generate the
document using the xml.dom.minidom APIs. This slightly changes the
formatting, so the tests are adjusted for that.

In addition, we can now easily overload the make function in the SQL
base class so that the children are always generated the same. This also
means we no longer need the _make_children function.