Improve Transport Exceptions
Add TransportError as a base class for all transport errors. This now
becomes the base class of the existing TransportClosedException. This
exception is also renamed to TransportClosedError, since that fits
better with the standard Python exception naming scheme, but a
deprecated alias is kept for compatibility.

In addition, add these classes to the docs and fix the doc strings so
the names are properly resolved and linkified.

Fixes #60