Use logging module instead of custom tracing
Tracing now sends INFO level log messages to the "itoolkit-trace"
logger instead of writing to stdout or a file. This gives more
flexibility and can be simulated similarly with built-in Python log
handlers. A doc page is added which documents how to set up a log
handler to capture the logs equivalently to the old trace_open and
trace_close mechanism.

The existing trace_open and trace_close are now marked deprecated. They
are now just wrappers around the Python logging infrastructure and add a
StreamHandler or FileHandler to simulate the old behavior. The
trace_write and trace_hexdump methods were removed as they were really
internal methods and likely nobody was using them anyway.

Fixes #61