ci: Fix build action
With 33621cd, now all builds are failing. This seems to be due to
the Python dependency version checks. Improve these so they're not
overlapping and the 3.11 check actually includes 3.11.

Even still Python 3.11 fails in GitHub Actions with only
"Error: Process completed with exit code 1.". For now, just disable this
version. In the process of diagnosing, it seems setup-python@v2 uses a
deprecated feature and was giving warnings so update to the latest
version. This does not fix the above error, unfortunately.