ci: Handle Poetry install on Python 2.7 and 3.4
Poetry complains if our dependencies can't be resolved for all versions
of Python we claim to support. Our dev dependencies include pytest and
we claim to support Python 2.7, 3.4, and 3.5 so we also need a
compatible version of pytest for those Python releases. pytest 4.6 is
the last to support these releases as per
https://docs.pytest.org/en/7.0.x/py27-py34-deprecation.html